+++ echelon/actions/check-username.php
@@ -1,7 +1,7 @@
 <?php
 // plubic page keep things simple
 require '../inc/functions.php';
-require '../inc/config.php';
+require '../app/config.php';
 require '../app/classes/LegacyDatabase.php'; // require sessions class
 
 // set and clean var