+++ echelon/banlist.php
@@ -3,7 +3,7 @@ error_reporting(E_ALL ^ E_NOTICE); // show all errors but notices
 $page = 'banlist';
 
 require_once 'inc/ctracker.php'; // anti worm injection protection
-require_once 'inc/config.php'; // load the config file
+require_once 'app/config.php'; // load the config file
 
 if(INSTALLED != 'yes') // if echelon is not install (a constant is added to the end of the cnfig during install) then die and tell the user to go  install Echelon
 	die('You still need to install Echelon. <a href="install/index.php">Install</a>');